+++ hypryou/utils_cy/levenshtein.pyi
@@ -2,9 +2,9 @@ def levenshtein_distance(s1: str, s2: str) -> int:
     ...
 
 
-def compute_score(s1: str, s2: str) -> int:
+def compute_score(s1: str, s2: str) -> float:
     ...
 
 
-def compute_text_match_score(s1: str, s2: str) -> int:
+def compute_text_match_score(s1: str, s2: str) -> float:
     ...
